tweak: add watch into parameters of useContractQuery
